Consolidate start_container test stub into existing harness
Problem: container_setup_test.py reimplemented a recording _StubOuter +
_outer() factory and a base64-decode helper, plus two start_container
behavioral tests, all of which duplicate the existing _StubOuter / _outer /
_stub harness in _outer_helpers_test.py (which already tests start_container).
Fix: move the start_container behavioral tests into _outer_helpers_test.py,
reusing its stub harness and a new shared _decode_remote_sh_command helper;
reduce container_setup_test.py to the pure-function tests (script shape,
shell-quoting, sh -n syntax check, base64 round-trip) that need no stub.
